requirement.txt
#to do
transfomers
pytorch
torchvision
detectron2
cv2
matplotlib
numpy
fastai